Add NAT detection cathedral side.
The NAT detection is done by having peers send a NATDETEK message
to the cathedral_nat_port when sending a KATEDRAL message to
the normal default port.

The cathedral will use this information to detect symmetric NAT
and will automatically hand out the correct information to the
peers for p2p discovery or not if it decides that relaying is
the best option.