Small improvement.
In both bless and confess call nyfe_zeroize() before exiting
on the two data structures that were registered when the
processes started.

Note that these both were explicitly wiped already but if you
call nyfe_zeroize_register(), it should be followed by a nyfe_zeroize().

So be a good citizen.