Restructure chapel_offer_encrypt() a bit.
Do the sanctum_cipher_kdf() call first, that way in case it
fails we don't have to worry about a copy of the session key
lingering into key (which points to op->data.offer.key).

And always go to cleanup, that way the offer recreation
triggers after the TTL expires in case of sanctum_cipher_kdf()
failures.