Add traffic analysis protection.
This comes in the form of "encapsulation" where the purgatory processes
will encapsulation / decapsulate the relevant packets with yet another
ESP header and 64-bit packet number (as per RFC 4106) which are filled
in with correct looking data as traffic is flowing.

The encapsulation thus comes with an overhead of 32 bytes.

The inner ESP header and 64-bit packet number are XOR'd with a mask
generated as follows:

mask = KMAC256(traffic_encap_key, fake_hdr + 128-bit random value)

This in turn hides all sanctum related magic SPI/SEQ combinations that
sanctum uses to denote key exchanges, NAT detection or cathedral
messages and makes the traffic blend in with other IPSec traffic.

Note that only the magic values were really a worry from a traffic
analysis point of view as they were unique to sanctum.

When tfc is turned on in combination with this, the only work left to
be done is the key exchange and cathedral packets size adapations as
they're still fixed sizes.

Hymn also supports the "encap" keyword now when configuring tunnels.

Sanctum shall work as before if encapsulation is not configured.