Add support for tap devices on OpenBSD and Linux.
This allows the creation of tap devices on both the Linux
and OpenBSD platforms, allowing to transport of ethernet
frames over the sanctum tunnel.

Works in all expected configurations:
	normal mode, pilgrim/shrine mode or when relaying over a cathedral.

When specifying the use of tap, a tunnel config setting is optional and
can be ignored, but you can still use it to configure ipv4 on the
tap interface if so desired.

The heaven-tx process will filter on etherType and allows
arp, vlan tagged traffic, ipv4 traffic and ipv6 traffic.

This opens up for some interesting ways of using sanctum.