cathedral: Let cathedral detect changed settings.
Before you needed to send a SIGHUP to the cathedral process for it
to reload the settings file but thats honestly just wonky for
external processes trying to signal rewrites.

Instead, let the cathedral check if the settings file its
last modified timestamp changed and reload it if thats the
case.

This happens every 10 seconds.