Rewrite how the cathedral looks up tunnels.
With larger setup it was essentially flawed and could select the
incorrect tunnel leading to connectivity issues when federation
between multiple cathedrals.

This commit changes that. The sanctum instances will now announce
to the cathedral their RX SPIs so that the cathedral can lookup
peers based on those.

For key offerings the cathedral now looks at the flock tied into
the offer that was sent instead of attempting to match everything
based on ip:port.