Allow p2p discovery across *all* cathedrals.
When peer information was being federated we would lose the
relevant information for a cathedral to notify one of its
clients about its peers NAT capabilities or address.

So if client A was connected to cathedral 1 and client B was connected
to cathedral 2 traffic would always flow as follows:

A <---> 1 <---> 2 <---> B

even if both peers would be p2p capable.

This is not ideal as in large deployments of cathedrals you do not
want only cathedral local clients to move to p2p.

This commit changes that and allows for the required information
to be federated across all cathedrals in the infrastructure.