cathedral: Small improvements to peer-info.
Let sanctum clients carry their 64-bit instance ID that is generated
at startup in the info messages. This allows the cathedral to detect
if a restart happened and allows it to reset its peerinfo state to 0.

For peers sharing the same external ip, send the cathedral address
instead. The cathedral cannot determine if two peers are on the same
internal net when coming from the same external ip, and thus we need
to force a relay.