Add the remembrance.
This feature allows sanctum instances to obtain federation
information from its configured cathedral, essentially
allowing the cathedrals to distribute their knowledge
of other cathedrals in the same network.

If the current cathedral becomes unresponsive, the sanctum instance
will automatically select the next cathedral in the remembrance
list. This provides automatic failover between cathedrals for
sanctum instances in the event a cathedral goes away.

In order to turn on remembrance, you must configure your
sanctum instance with cathedral_remembrance <path> where
the given path is the file where the instance will save
the remembrance list to (in case of restart).