Rework erasing keys logic a bit.
After the changes that introduced the asymmetry we no longer install
RX keys immediately and thus can have a situation where we have
an active offer and no pending RX key when sanctum tries to
erase the keys due to inactivity.

If this is the case, chapel would deadlock on attempting to
instruct confess to erase keys as the io->rx state would
be ERASE and won't be cleared due to a mismatch in SPIs.

We fix this by generalizing erasing of keys so it includes
both the active RX and potentially pending RX.

We also make sure we wakeup the bless/confess processes
before and after installing new keys for them to pickup.

Additionally, we do a burst of 5 heartbeats, one per second
when we install new TX keys.