Change ML-KEM to ML-KEM-1024.
This requires us to do offer splitting for KEM offers. Right now
each offer message is 589 bytes, which is fine. So we send 4 packets
per PK or CT message.

We need to increase the ring queues for chapel and offer to facility
this new behaviour. Also fix clearing the wrong stat struct in confess.

Remove the clearing of the DF bit again as we do not need this
anymore in this case. The packets offering packets are small
enough.