Improvements to the key exchange.
We currently derive a traffic_base key from the shared secret.
This traffic_base is used as part of the IKM for both session
key KDFs (RX/TX).

The x25519 and mlkem1024 inputs are already unique per direction, so
continue this trend and instead of traffic_base derive two unique
keys from the shared secret that are used as the symmetrical inputs
for the RX and TX KDFs respectively.