chapel: Fix a rare PRECOND hit.
Fix a rare race where we could PRECOND in sanctum_install_key_material()
due to the fact that offer->remote.spi was not yet set. This was the case
if we happened to receive the CT message before the peer its PK message.

Due to the nature of internet this can very much occur so we handle it.