liturgy: Add a way of signaling peers.
Normally, liturgy mode means autodiscovery for all peers that are
online in your flock and this is great, but we can use this to
also signal specific peers which opens up an entire other world
of possibilities.

We keep the default behaviour for liturgy but allow the new behaviour
to work in tandem with it.