More startup improvements.
Obviously we cannot move sanctum_proc_start() before daemon()
in case we're backgrounding so fix that change.

Remove the early status for logging and just log to syslog
when we're in going to daemonize even if we're still in
early startup phase.

Change hymn to redirect the user to syslog in case of a failure.