Introduce flock domains in cathedrals.
A cathedral identifies peers as belonging to specific flocks. These are
64-bit random integers that uniquely identify a group of up to 256 peers.

With this commit we introduce the ability to create separate domains
inside of these flocks. These peers in these different domains are
cryptographically separated by generating different key material
for the offer_base_key and traffic_base_key keys depending on the domain.

This helps in a scenario where you have a device and would like to use
different applications that use sanctum as the transport (confessions, litany)
but do not want to create new flocks and issue new keys for these different
applications - which is the case today.