cathedral: Introduce xflock.
The xflock feature allows sanctum instances that live in different
cathedral flocks to establish tunnels with each other.

Each xflock will get its own ambry that must be generated specifically
for that xflock group.

Each xflock must be explicitly configured in a cathedral config.

This required a protocol change and will break existing setups.

Note that with the protocol changes more meta-data is carried
in the clear so having an encapsulation secret for your setup
is a good idea.