Change so we no longer use the KEK directly.
Before we used the KEK directly when unwrapping Ambry entries, we change
this so that we derive a key instead using the same sanctum_base_key()
derivation function as everything else.

This separates the KEK keys based on src:dst flocks giving us different
keys for different flocks (internal / external).

Change the ambry tool to generate the base KEKs from which the internal
flock ones are derived.

Change the ambry tool to add an export command that can be used to
derive new KEKs for a specific xflock purpose, these could for
example be sent to your xflock partner so they can generate
xflock ambries, or vise versa.