Introduce cathedral signing keys (COSK).
These are used to authenticate offers sent from clients to cathedrals.

Previously the CS (cathedral secret) was used for both confidentiality
and integrity of any offers sent and received from and to cathedrals,
meaning an attacker who compromised a cathedral could store these
secrets for later and mount an attack at any time from anywhere
allowing disruption of client traffic.

With the introduction of COSK this however is no longer possible
as signatures are now required for all offers sent from clients
to cathedrals and the signatures are kept intact when federating
to other cathedrals, allowing other cathedrals to also verify
the authenticity of these.

This is a breaking change, anyone using cathedrals will need to adjust
for the introduction of COSKs and generate COSKs for their devices and
place the public keys on their cathedrals.

This is also one of the final big changes before a 1.0 release.