Split up KDF label for offers.
Before we used the same KDF label for offers sent to both
peers and cathedrals. While there is nothing wrong with
this it is good hygiene to split these up as the keys
are used for different purposes.

We also split up the KDF prefix compile time overriding
so that it is possible to override the traffic KDF labels
while retaining the cathedral ones, allowing domain separation
for different sanctum purposes but still able to use same
cathedral infrastructure.

This is a breaking change, any infra will need to be rebuilt.