chapel: Fix a very rare desync in key exchanges.
If one peer is restarting quickly we could get into a desynced state
where we are unable to recover from and the exchanges just keep going
until one peer is restarted.

This commit fixes that by simply clearing the exchange and forcing
a new one to be scheduled for a re-sync.